Adds Vector Length
